13: Part 2
Quite dissatisfied with this. I stumbled across Chinese Remainder
Theorem, which apparently a lot of people "just know", but I've never
heard of it and I had to just grab an algorithm for that off the 'net.

Independently I got to the point of a bunch of simultaneous equations,
but I end up with N-1 equations in N unknowns. Wolfram was able to solve
it but I'm not sure how.

The "increase step size by product" approach (maybe called seiving?) I
saw in some threads related to this challenge - so that's definitely
cheating.