19: Part2 "the hard way"
This looks simple now, but I fought with it for hours.
Memoization is essential to get reasonable runtime out of this. I think
the bug I spent all day chasing down was just the "optionIsBad" case.
I did just 'break' in that case, but that break is nested one level too
deep in the loops, so it wasn't doing the right thing.

Anyway, I think this is probably the "hard" way alluded to on the puzzle
page (at least, I hope so!).

I couldn't really figure out the "simple" way, at least not with the way
I did Part 1. I imagine, the "intended" solution to Part 1 is to work
through all the rules and replace them with the concrete substrings they
represent, then match those against the message. Presumably, for Part 2
you'd then add a couple of exceptions, working from the front and the
back, then attempting to replace the middle with one of the
recursive/repeat rules.