23: WIP: Part 2
An implementation that I think would work, but it's too slow (would take
~1 hour), so presumably there's a better way.

Maybe some kind of hash-map of indices? But then wouldn't "insert" be
updating ~1m indices (or 500k on average) - that sounds slow.

I wondered if some kind of segment approach with a base, increment and
number-of-values-in-segment would help, but the list is pretty well
fragmented by the time we reach 1m moves, so I'd expect the segments to
be very small (tends towards segment size of 1, assuming the algorithm
is doing a decent shuffle).